change ido default buffer/file methods
Ido has a nasty feature of switching to frame in which window with
file/buffer in question is already opened, instead opening it in current
window. Most probably it was designed for lesser, non-tiling WMs.

Patch fixes this behaviour, forcing C-x C-f and C-x b to open
file/buffer in the current window.